--Empty Message--

+++ 2019/prac-2017-r1c/A.cc
@@ -55,7 +55,7 @@ void process() {
     double sum = M_PI * sqr((LL)data[i-1].first);
     for(int j=i-K;j<i;j++) {
       cerr << data[j].first << " ";
-      sum += 2 * M_PI * data[j].first * data[j].second;
+      sum += 2 * M_PI * (double)data[j].first * data[j].second;
     }
     cerr << endl;
     if(res < sum) res = sum;
